factor out remove/apply style mechanisms into manage-style.js
